django-easyvisa/requirements.txt
asgiref==3.2.9
certifi==2020.4.5.2
chardet==3.0.4
DAWG-Python==0.7.2
Django==3.0.7
django-admin-sortable2==0.7.6
django-appconf==1.0.4
django-autoslug==1.9.7
django-ckeditor==5.9.0
django-cleanup==5.0.0
django-imagekit==4.0.2
django-ipware==2.1.0
django-js-asset==1.2.2
django-nested-inline==0.4.2
djangorestframework==3.11.0
docopt==0.6.2
geoip2==3.0.0
idna==2.9
maxminddb==1.5.4
pilkit==2.0
Pillow==7.1.2
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
python-decouple==3.3
pytils==0.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
